fix(migrations): change active column from SMALLINT to INTEGER on Postgres
The sqlx Any backend maps SMALLINT inconsistently across backends when
decoding into i64.  Using INTEGER (which maps to i64 reliably on all
three backends) avoids potential type-mismatch panics at runtime.

+++ migrations/postgres/0008_operators.sql
@@ -8,6 +8,6 @@ CREATE TABLE operators (
     gssapi_principal TEXT        UNIQUE,       -- Kerberos principal; NULL = no GSSAPI auth
     created_at       TEXT        NOT NULL,     -- RFC 3339
     last_seen_at     TEXT,                     -- RFC 3339
-    active           SMALLINT    NOT NULL DEFAULT 1,
+    active           INTEGER     NOT NULL DEFAULT 1,
     CHECK(cert_fingerprint IS NOT NULL OR gssapi_principal IS NOT NULL)
 );
